# docs/openapi.yaml
swagger: '2.0'
basePath: /
consumes:
- application/json
definitions:
  container-image:
    properties:
      name:
        description: The container image url
        type: string
    type: object
...
  data-filename:
    properties:
      name:
        description: The name of the data file
        type: string
    type: object
...
  executionoptiondefaults:
    properties:
      variables:
        default: []
        description: The default values of the variables in the virtual experiment
          DSL
        items:
          $ref: '#/definitions/variablewithdefaultvalues'
        type: array
    type: object
...
  input-filename:
    properties:
      name:
        description: The name of the input file
        type: string
    type: object
  internal-experiment:
    properties:
      pvep:
        allOf:
        - $ref: '#/definitions/ve'
        description: The Parameterised Virtual Experiment Package that parameterizes
          the execution of the workflow
      workflow:
        allOf:
        - $ref: '#/definitions/internal-experiment-workflow'
        description: The definition of the workflow
    type: object
  internal-experiment-workflow:
    properties:
      dsl:
        description: The dictionary representing an experiment in DSL format
        type: object
    type: object
...
  option-valuefrom-secretkeyref:
    properties:
      key:
        description: Name of the key inside the Kubernetes Secret object
        type: string
      name:
        description: Name of the Kubernetes Secret object
        type: string
    required:
    - name
    type: object
  orchestratorresources:
    properties:
      cpu:
        description: How many cores to request for the orchestrator executing the
          workflow
        type: string
      memory:
        description: How much memory to request for the orchestrator executing the
          workflow
        type: string
    type: object
...
  query-experiment-common:
    properties:
      matchPackageVersion:
        default: false
        description: Whether to match the version of packages
        type: boolean
      mustHaveOnePackage:
        default: true
        description: Match only parameterised virtual experiment packages with just
          one base package
        type: boolean
    type: object
  query-experiments:
    properties:
      common:
        $ref: '#/definitions/query-experiment-common'
      package:
        $ref: '#/definitions/query-experiments-package'
      relationship:
        $ref: '#/definitions/query-experiments-relationship'
    type: object
  query-experiments-package:
    properties:
      definition:
        $ref: '#/definitions/ve-base-package'
    type: object
  query-experiments-relationship:
    properties:
      identifier:
        description: The identifier of the relationship
        type: string
      transform:
        $ref: '#/definitions/query-experiments-relationship-transform'
    type: object
  query-experiments-relationship-transform:
    properties:
      matchInputGraph:
        default: false
        description: Whether to query using the package of the inputGraph
        type: boolean
      matchOutputGraph:
        default: false
        description: Whether to query using the package of the inputGraph
        type: boolean
    type: object
...
  valueinplatform:
    properties:
      platform:
        type: string
      value:
        type: string
    type: object
  variablewithdefaultvalues:
    properties:
      name:
        type: string
      valueFrom:
        default: []
        items:
          $ref: '#/definitions/valueinplatform'
        type: array
    type: object
  ve:
    properties:
      base:
        allOf:
        - $ref: '#/definitions/ve-base'
        description: The configuration of the base packages
      metadata:
        allOf:
        - $ref: '#/definitions/ve-metadata'
        description: Metadata of the parameterised virtual experiment package
      parameterisation:
        allOf:
        - $ref: '#/definitions/ve-parameterisation'
        description: The parameterisation options
    required:
    - base
    - metadata
    type: object
  ve-base:
    properties:
      packages:
        default: []
        description: The packages that make up this parameterised virtual experiment
          package
        items:
          $ref: '#/definitions/ve-base-package'
        type: array
    required:
    - packages
    type: object
  ve-base-package:
    properties:
      config:
        allOf:
        - $ref: '#/definitions/ve-base-package-config'
        description: Configuration options for the package
      name:
        default: main
        description: Unique name of base package in this virtual experiment entry.
          Defaults to "main"
        type: string
      source:
        allOf:
        - $ref: '#/definitions/ve-base-package-source'
        description: Information on the location of the package
    required:
    - source
    type: object
  ve-base-package-config:
    properties:
      manifestPath:
        description: The path to the manifest file in the package
        type: string
      path:
        description: The path to the workflow definition in the package
        type: string
    type: object
  ve-base-package-source:
    properties:
      dataset:
        allOf:
        - $ref: '#/definitions/ve-base-package-source-dataset'
        description: The configuration for a package that exists on a Dataset. Mutually
          exclusive with @git
      git:
        allOf:
        - $ref: '#/definitions/ve-base-package-source-git'
        description: The configuration for a package that exists on a git server.
          Mutually exclusive with @dataset
    type: object
  ve-base-package-source-dataset:
    properties:
      location:
        allOf:
        - $ref: '#/definitions/ve-base-package-source-dataset-info'
        description: The Dataset which holds the package
      security:
        allOf:
        - $ref: '#/definitions/ve-base-package-source-dataset-info'
        description: The information required to get the package from the dataset
      version:
        description: The version of the package
        type: string
    required:
    - location
    type: object
  ve-base-package-source-dataset-info:
    properties:
      dataset:
        description: The name of the dataset
        type: string
    required:
    - dataset
    type: object
  ve-base-package-source-git:
    properties:
      location:
        allOf:
        - $ref: '#/definitions/ve-base-package-source-git-location'
        description: The location of the package on git
      security:
        allOf:
        - $ref: '#/definitions/ve-base-package-source-git-security'
        description: The information required to get the package from git
      version:
        description: The commit id of the package on git
        type: string
    required:
    - location
    type: object
  ve-base-package-source-git-location:
    properties:
      branch:
        description: Git branch name, mutually exclusive with @tag and @commit
        type: string
      commit:
        description: Git commit digest, mutually exclusive with @branch and @tag
        type: string
      tag:
        description: Git tag name, mutually exclusive with @branch and @commit
        type: string
      url:
        description: Git url, must provide this if package is hosted on a Git server
        type: string
    required:
    - url
    type: object
  ve-base-package-source-git-security:
    properties:
      oauth:
        allOf:
        - $ref: '#/definitions/ve-base-package-source-git-security-oauth'
        description: The oauth-token to use when retrieving the package from git
    type: object
  ve-base-package-source-git-security-oauth:
    properties:
      value:
        description: The value of the oauth-token, when using this field the runtime
          service will store the token in a new Kubernetes Secret and update the Parameterised
          Virtual Experiment Package to reference the Secret instead of the oauth-token
          directly. Mutually exclusive with @valueFrom
        type: string
      valueFrom:
        allOf:
        - $ref: '#/definitions/ve-base-package-source-git-security-oauth-valuefrom'
        description: A pointer to the oauth-token. Mutually exclusive with @value
    type: object
  ve-base-package-source-git-security-oauth-valuefrom:
    properties:
      secretKeyRef:
        allOf:
        - $ref: '#/definitions/option-valuefrom-secretkeyref'
        description: Description of the Kubernetes Secret key that contains the value
          of the oauth-token
    type: object
  ve-metadata:
    properties:
      package:
        allOf:
        - $ref: '#/definitions/ve-metadata-package'
        description: Metadata aboud the parameterised virtual experiment package
      registry:
        allOf:
        - $ref: '#/definitions/ve-metadata-registry'
        description: Metadata that the registry generates
    required:
    - package
    type: object
  ve-metadata-package:
    properties:
      description:
        description: The description of the parameterised virtual experiment package
        type: string
      keywords:
        default: []
        description: Keywords associated with the parameterised virtual experiment
          package
        items:
          type: string
        type: array
      license:
        description: The license of the parameterised virtual experiment package
        type: string
      maintainer:
        description: The maintainer of the parameterised virtual experiment package
        type: string
      name:
        description: The name of the parameterised virtual experiment package
        type: string
      tags:
        default: []
        description: The tags associated with the parameterised virtual experiment
          package
        items:
          type: string
        type: array
    required:
    - name
    type: object
  ve-metadata-registry:
    properties:
      containerImages:
        default: []
        items:
          $ref: '#/definitions/container-image'
        type: array
      createdOn:
        type: string
      data:
        items:
          $ref: '#/definitions/data-filename'
        type: array
      digest:
        type: string
      executionOptionsDefaults:
        $ref: '#/definitions/executionoptiondefaults'
      inputs:
        default: []
        items:
          $ref: '#/definitions/input-filename'
        type: array
      interface:
        default: {}
        type: object
      tags:
        default: []
        items:
          type: string
        type: array
      timesExecuted:
        default: 0
        type: integer
    type: object
  ve-parameterisation:
    properties:
      executionOptions:
        allOf:
        - $ref: '#/definitions/ve-parameterisation-executionoptions'
        description: Parameterisation options for settings that users can override
          at execution time (within constraints)
      presets:
        allOf:
        - $ref: '#/definitions/ve-parameterisation-presets'
        description: Parameterisation options for settings that users cannot override
          at execution time
    type: object
  ve-parameterisation-executionoptions:
    properties:
      data:
        default: []
        description: Parameterisation options for data files which the users can override
          at execution time
        items:
          $ref: '#/definitions/data-filename'
        type: array
      platform:
        default: []
        description: Parameterisation options for the platform to specialize the workflow
          at execution time.The users can override this option at execution time (within
          constraints)
        items:
          type: string
        type: array
      variables:
        default: []
        description: Parameterisation options for variables which the users can override
          at execution time (within constraints)
        items:
          $ref: '#/definitions/ve-parameterisation-executionoptions-variable'
        type: array
    type: object
  ve-parameterisation-executionoptions-variable:
    properties:
      name:
        type: string
      value:
        description: This is the default value of the variable, providing this field
          means that the variable can recieve *any* value
        type: string
      valueFrom:
        description: An array of choices that this variable must be set to. If at
          execution time the variable does not receive a value, then the default value
          is the first value in this array
        items:
          $ref: '#/definitions/ve-parameterisation-executionoptions-variable-choice'
        type: array
    type: object
  ve-parameterisation-executionoptions-variable-choice:
    properties:
      value:
        description: The value of the variable, must be compatible with the parameterisation
          options of the parameterised virtual experiment package
        type: string
    type: object
  ve-parameterisation-presets:
    properties:
      data:
        default: []
        description: The configuration for data-files
        items:
          $ref: '#/definitions/data-filename'
        type: array
      environmentVariables:
        default: []
        description: Environment variables to inject into the processes that orchestrate
          the execution of the workflow
        items:
          $ref: '#/definitions/ve-parameterisation-presets-environment-variables'
        type: array
      platform:
        description: The platform to specialize the workflow for
        type: string
      runtime:
        allOf:
        - $ref: '#/definitions/ve-parameterisation-runtime'
        description: The runtime configuration of the orchestrator executing the workflow
      variables:
        default: []
        description: The preset variables
        items:
          $ref: '#/definitions/ve-parameterisation-presets-variable'
        type: array
    type: object
  ve-parameterisation-presets-environment-variables:
    properties:
      name:
        description: Name of the environment variable to inject into the containers
          in the orchestrator pod
        type: string
      value:
        description: The value of the environment variable
        type: string
    required:
    - name
    - value
    type: object
  ve-parameterisation-presets-variable:
    properties:
      name:
        description: The name of the variable
        type: string
      value:
        description: The preset value of the variable
        type: string
    required:
    - name
    - value
    type: object
  ve-parameterisation-runtime:
    properties:
      args:
        default: []
        description: Commandline arguments to the orchestrator executing the workflow
        items:
          type: string
        type: array
      resources:
        allOf:
        - $ref: '#/definitions/orchestratorresources'
        description: Resource requests for the orchestrator executing the workflow
    type: object
...
info:
  description: Launch/Monitor/Stop experiments
  title: Runtime Service of the Simulation Toolkit for Scientific Discovery (ST4SD)
  version: '1.0'
paths:
...
  /experiments/:
    get:
      operationId: get_experiment_list
      parameters:
      - default: json
        description: Output format
        enum:
        - json
        - python
        - python-pretty
        in: query
        name: outputFormat
        type: string
      - default: n
        description: Whether to hide hte registry metadata or not
        enum:
        - y
        - n
        in: query
        name: hideMetadataRegistry
        type: string
      - default: y
        description: Whether to hide fields whose value is None
        enum:
        - y
        - n
        in: query
        name: hideNone
        type: string
      - default: y
        description: Whether to hide Beta fields
        enum:
        - y
        - n
        in: query
        name: hideBeta
        type: string
      responses:
        '200':
          description: Success
      summary: List all experiments
      tags:
      - experiments
    post:
      operationId: post_experiment_list
      parameters:
      - in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/ve'
      responses:
        '200':
          description: OK
        '400':
          description: Experiment ID is invalid
        '409':
          description: Experiment ID already exists
      summary: Add an experiment
      tags:
      - experiments
...
  /experiments/{identifier}:
    delete:
      operationId: delete_experiment
      responses:
        '200':
          description: Success
        '404':
          description: Unknown experiment
        '500':
          description: Internal error
      summary: Delete an experiment
      tags:
      - experiments
    get:
      operationId: get_experiment
      parameters:
      - default: json
        description: Output format
        enum:
        - json
        - python
        - python-pretty
        in: query
        name: outputFormat
        type: string
      - default: n
        description: Whether to hide hte registry metadata or not
        enum:
        - y
        - n
        in: query
        name: hideMetadataRegistry
        type: string
      - default: y
        description: Whether to hide fields whose value is None
        enum:
        - y
        - n
        in: query
        name: hideNone
        type: string
      - default: y
        description: Whether to hide Beta fields
        enum:
        - y
        - n
        in: query
        name: hideBeta
        type: string
      responses:
        '404':
          description: Unknown experiment
      summary: Fetch an experiment given its identifier
      tags:
      - experiments
    parameters:
    - description: The package identifier. It must contain a $packageName and may
        include either a tag suffix (`:${tag}`) or a digest suffix (`@${digest}`).
        If both suffixes are missing then the identifier implies the `:latest` tag
        suffix.
      in: path
      name: identifier
      required: true
      type: string
...
  /internal-experiments/:
    post:
      operationId: post_internal_experiments
      parameters:
      - in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/internal-experiment'
      responses:
        '200':
          description: Success
      tags:
      - internal-experiments
  /query/experiments/:
    post:
      operationId: post_query_experiments
      parameters:
      - in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/query-experiments'
      responses:
        '200':
          description: Success
      tags:
      - query
...
produces:
- application/json
...